refactor: extract ios discovery helpers
